Adding user state
