+++ Web/utkeyrslukerfi-web/src/constants/index.jsx
@@ -3,6 +3,7 @@ export const CREATE_USER = 'CREATE_USER'
 export const GET_DELIVERIES = 'GET_DELIVERIES'
 export const GET_DELIVERY = 'GET_DELIVERY'
 export const UPDATE_DELIVERY = 'UPDATE_DELIVERY'
+export const SET_USER = 'SET_USER'
 // URLS
 export const USER_URL = 'https://localhost:5001/api/users'
 export const DELIVERY_URL = 'https://localhost:5001/api/deliveries'
