+++ Web/utkeyrslukerfi-web/src/reducers/index.jsx
@@ -1,8 +1,10 @@
 import { combineReducers } from 'redux'
-import users from './userReducer'
+import users from './usersReducer'
 import deliveries from './deliveryReducer'
+import user from './userReducer'
 
 export default combineReducers({
+  user,
   users, 
-  deliveries
+  deliveries,
 })
