+++ Web/utkeyrslukerfi-web/src/reducers/userReducer.jsx
@@ -1,9 +1,8 @@
-import { GET_USERS, CREATE_USER } from '../constants'
+import { SET_USER } from '../constants'
 
-const userReducer = (state = [], action) => {
+const userReducer = (state = {}, action) => {
   switch (action.type) {
-    case GET_USERS: return action.payload
-    case CREATE_USER: return [...state, action.payload]
+    case SET_USER: return action.payload
     default: return state
   }
 }
