+++ Web/utkeyrslukerfi-web/src/views/UpdatePasswordForm/index.jsx
@@ -4,7 +4,7 @@ import Form from 'react-bootstrap/Form'
 import Col from 'react-bootstrap/Col'
 import Row from 'react-bootstrap/Row'
 import Button from 'react-bootstrap/Button'
-import { updatePassword } from '../../actions/userActions'
+import { updatePassword } from '../../actions/usersActions'
 import { connect } from 'react-redux'
 
 const UpdatePasswordForm = ({ updatePassword }) => {
